+++ internal/provider/alerts/alert_schema/v1.go
@@ -198,11 +198,7 @@ func V1() schema.Schema {
 					"logs_anomaly": schema.SingleNestedAttribute{
 						Optional: true,
 						Attributes: map[string]schema.Attribute{
-							"custom_evaluation_delay": 	   evaluationDelaySchema(),
-							"percentage_of_deviation":     schema.Float64Attribute{
-								Optional:            true,
-								MarkdownDescription: "The percentage of deviation from the baseline for triggering the alert.",
-							},
+							"custom_evaluation_delay":     evaluationDelaySchema(),
 							"logs_filter":                 logsFilterSchema(),
 							"notification_payload_filter": notificationPayloadFilterSchema(),
 							"rules": schema.SetNestedAttribute{
@@ -417,12 +413,8 @@ func V1() schema.Schema {
 					"metric_anomaly": schema.SingleNestedAttribute{
 						Optional: true,
 						Attributes: map[string]schema.Attribute{
-							"custom_evaluation_delay": 	evaluationDelaySchema(),
-							"percentage_of_deviation": 	schema.Float64Attribute{
-								Optional:            true,
-								MarkdownDescription: "The percentage of deviation from the baseline for triggering the alert.",
-							},
-							"metric_filter": 			metricFilterSchema(),
+							"custom_evaluation_delay": evaluationDelaySchema(),
+							"metric_filter":           metricFilterSchema(),
 							"rules": schema.SetNestedAttribute{
 								Required:   true,
 								Validators: []validator.Set{setvalidator.SizeAtLeast(1)},